requirements.txt
absl-py==1.4.0
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
branca==0.6.0
cachetools==5.3.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.1.1
distlib==0.3.7
Django==4.1.3
executing==1.2.0
filelock==3.13.1
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.47.2
fsspec==2023.12.2
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.6
imutils==0.5.4
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.3
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
Pygments==2.16.1
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
Python-IO==0.3
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.12.0
seaborn==0.13.1
setuptools>=69.0.3
six==1.16.0
sklearn==0.0.post7
smmap==5.0.1
SpeechRecognition==3.10.0
sphinx-rtd-theme==0.2.4
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.4
ultralytics==8.1.3
urllib3==1.26.18
virtualenv==20.24.5
wcwidth==0.2.6
webcolors==1.13
Werkzeug==2.3.6
wordcloud==1.9.2
wrapt==1.15.0
